Spreaded main class into main.rs and dragons.rs.
