+++ src/bin/main.rs
@@ -11,14 +11,6 @@ extern crate stm32f7_discovery;
 #[macro_use]
 extern crate alloc;
 
-const IMG: [u8; 30 * 30 * 2] = *include_bytes!("dragonResized.data");
-const COLORS: [(u8, u8, u8); 5] = [
-    (255, 0, 0),
-    (0, 128, 0),
-    (0, 0, 128),
-    (192, 192, 192),
-    (255, 255, 0),
-];
 const max_simultaneous_dragons_on_screen: u8 = 15;
 
 use alloc_cortex_m::CortexMHeap;
@@ -36,6 +28,10 @@ use stm32f7_discovery::{
     system_clock::{self, Hz},
     touch,
 };
+
+mod dragons;
+use dragons::Box;
+use dragons::Vector2d;
 
 #[global_allocator]
 static ALLOCATOR: CortexMHeap = CortexMHeap::empty();
@@ -118,23 +114,15 @@ fn main() -> ! {
 
     let mut number_of_dragons = 0;
     let mut last_dragon_create = system_clock::ticks();
-    let mut rand = rand::rngs::StdRng::seed_from_u64(318273678346);
+    // let mut rand = rand::rngs::StdRng::seed_from_u64(318273678346);
     let mut boxes: alloc::vec::Vec<Box> = vec![];
     let mut counter = 5;
     let mut last_led_toggle = system_clock::ticks();
     let mut last_render = system_clock::ticks();
     let mut last_second = system_clock::ticks();
+            let mut rand = rand::rngs::StdRng::seed_from_u64(318273678346);
     while number_of_dragons <= 8 {
-        let tupel = COLORS[rand.gen_range(0, 4)];
-        let color = Color::rgb(tupel.0, tupel.1, tupel.2);
-        let dragon = Box::new(
-            30,
-            rand.gen_range(60, 400),
-            rand.gen_range(60, 200),
-            rand.gen_range(-10, 10),
-            rand.gen_range(-10, 10),
-            color,
-        );
+        let dragon = Box::random(&mut rand);
         let mut intersect = false;
         for d in &boxes.clone() {
             if d.intersect(&dragon.clone()) {
@@ -180,16 +168,7 @@ fn main() -> ! {
             if ticks - last_dragon_create >= 10 {
                 if number_of_dragons < max_simultaneous_dragons_on_screen {
                     //add dragon
-                    let tupel = COLORS[rand.gen_range(0, 4)];
-                    let color = Color::rgb(tupel.0, tupel.1, tupel.2);
-                    let dragon = Box::new(
-                        30,
-                        rand.gen_range(60, 400),
-                        rand.gen_range(60, 200),
-                        rand.gen_range(-10, 10),
-                        rand.gen_range(-10, 10),
-                        color,
-                    );
+                    let dragon = Box::random(&mut rand);
                     let mut intersect = false;
                     for d in &boxes.clone() {
                         if d.intersect(&dragon.clone()) {
@@ -307,136 +286,5 @@ fn main() -> ! {
         }
         counter = 5;
         seconds = 0;
-    }
-}
-
-#[derive(Copy, Clone)]
-struct Vector2d {
-    x: i16,
-    y: i16,
-}
-#[derive(Copy, Clone)]
-struct Box {
-    size: u16,
-    pos: Vector2d,
-    vel: Vector2d,
-    col: Color,
-}
-
-impl Box {
-    fn new(size: u16, x: i16, y: i16, vel_x: i16, vel_y: i16, vel_col: Color) -> Self {
-        Self {
-            size,
-            pos: Vector2d { x: x, y: y },
-            vel: Vector2d { x: vel_x, y: vel_y },
-            col: vel_col,
-        }
-    }
-    fn render(
-        &mut self,
-        layer: &mut stm32f7_discovery::lcd::Layer<stm32f7_discovery::lcd::FramebufferArgb8888>,
-    ) {
-        for y in 0..self.size {
-            for x in 0..self.size {
-                let i = 20 + x as usize + self.pos.x as usize;
-                let j = 20 + y as usize + self.pos.y as usize;
-                let wert = IMG[2 * (x + y * 30) as usize + 1];
-                if wert > 25 {
-                    //let c = Color::rgb(255, 0, 0);
-                    layer.print_point_color_at(i, j, self.col)
-                }
-            }
-        }
-        // for i in (20 + self.pos.x)..=(20 + self.pos.x + self.size as i16) {
-        //     for j in (20 + self.pos.y)..=(20 + self.pos.y + self.size as i16) {
-        //         layer.print_point_color_at(i as usize, j as usize, Color::rgb(IMG[2*i as usize+j as usize], 0, 0))
-        //         //layer.print_point_color_at(i as usize, j as usize, self.col)
-        //     }
-        // }
-    }
-
-    fn derender(
-        &mut self,
-        layer: &mut stm32f7_discovery::lcd::Layer<stm32f7_discovery::lcd::FramebufferArgb8888>,
-        bg_color: Color,
-    ) {
-        for i in (20 + self.pos.x)..(20 + self.pos.x + self.size as i16) {
-            for j in (20 + self.pos.y)..(20 + self.pos.y + self.size as i16) {
-                layer.print_point_color_at(i as usize, j as usize, bg_color)
-            }
-        }
-    }
-
-    fn hit(&mut self, hit: &Vector2d) -> bool {
-        if self.pos.x + 20 <= hit.x && hit.x <= self.pos.x + self.size as i16 + 20 {
-            if self.pos.y + self.size as i16 + 20 >= hit.y && hit.y >= self.pos.y + 20 {
-                self.col = Color::from_hex(0xffffff);
-                return true;
-            }
-        }
-        return false;
-    }
-
-    fn collision(&self, b: &Box) -> Option<Vector2d> {
-        if self.intersect(b) {
-            return Some(Vector2d {
-                x: b.vel.x,
-                y: b.vel.y,
-            });
-        } else {
-            return None;
-        }
-    }
-
-    fn intersect(&self, b: &Box) -> bool {
-        if self.pos.x == b.pos.x
-            && self.pos.y == b.pos.y
-            && self.vel.x == b.vel.x
-            && self.vel.y == b.vel.y
-            && self.size == b.size
-        {
-            return false;
-        }
-        (math::fabsf(self.pos.x as f32 - b.pos.x as f32) * 2.0 < (self.size as f32 + b.size as f32))
-            && (math::fabsf(self.pos.y as f32 - b.pos.y as f32) * 2.0
-                < (self.size as f32 + b.size as f32))
-    }
-
-    fn update_speed(&mut self, new_speed: &Vector2d) {
-        self.vel = *new_speed;
-    }
-
-    fn next(&mut self) {
-        let HEIGHT = 252;
-        let WIDTH = 460;
-        let OFFSET = 20;
-        if OFFSET + self.size as i16 + self.pos.x as i16 + self.vel.x > WIDTH {
-            self.vel.x *= -1;
-        }
-        if OFFSET + self.pos.x as i16 + self.vel.x < 20 {
-            self.vel.x *= -1;
-        }
-        if OFFSET + self.size as i16 + self.pos.y as i16 + self.vel.y > HEIGHT {
-            self.vel.y *= -1;
-        }
-        if OFFSET + self.pos.y as i16 + self.vel.y < 20 {
-            self.vel.y *= -1;
-        }
-        self.pos.x += self.vel.x;
-        self.pos.y += self.vel.y;
-    }
-}
-
-impl PartialEq for Box {
-    fn eq(&self, b: &Box) -> bool {
-        if self.pos.x == b.pos.x
-            && self.pos.y == b.pos.y
-            && self.vel.x == b.vel.x
-            && self.vel.y == b.vel.y
-            && self.size == b.size
-        {
-            return true;
-        }
-        false
     }
 }
